notification

+++ frontend/src/views/Chefs/ChefListView.vue
@@ -2,6 +2,8 @@
 import { ref, reactive, onMounted } from 'vue'
 import chefService from '@/services/chef.js'
 import departementService from '@/services/departement.js'
+import Swal from 'sweetalert2'
+import { useToast } from 'vue-toastification'
 import {
   Users,
   CheckCircle2,
@@ -23,6 +25,7 @@ const departments = ref([])
 const toUpdate = ref(null)
 const search = ref('')
 const stats = ref([])
+const toast = useToast()
 
 // --- Formulaire Réactif ---
 const newChef = reactive({
@@ -98,7 +101,7 @@ const handleSubmit = async () => {
     await fetchData()
     resetForm()
   } catch (error) {
-    alert("Erreur lors de l'enregistrement")
+    toast.error("Erreur lors de l'enregistrement")
   }
 }
 
@@ -113,10 +116,21 @@ const openEditModal = (chef) => {
 }
 
 const handleDelete = async (id) => {
-  // if (confirm('Supprimer ce responsable ?')) {
-  await chefService.deleteChef(id)
-  fetchData()
-  // }
+  const result = await Swal.fire({
+    title: 'Supprimer ce responsable?',
+    text: 'Êtes-vous sûr de vouloir supprimer ce responsable de département?',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#dc2626',
+    cancelButtonColor: '#6b7280',
+    confirmButtonText: 'Oui, supprimer',
+    cancelButtonText: 'Annuler'
+  });
+
+  if (result.isConfirmed) {
+    await chefService.deleteChef(id)
+    fetchData()
+  }
 }
 
 onMounted(fetchData)
